pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lira-analysis"
version = "1.0.0"
description = "Implementation of LiRA (Likelihood Ratio Attack) membership inference attack under realistic assumptions"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Najeeb Jebreel", email = "najeeb.jebreel@urv.cat"}
]
keywords = [
    "membership inference",
    "privacy",
    "machine learning",
    "deep learning",
    "adversarial attacks",
    "LiRA"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "timm>=0.9.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.65.0",
]

[project.scripts]
lira = "utils.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/najeebjebreel/lira_analysis"
Documentation = "https://github.com/najeebjebreel/lira_analysis/blob/main/README.md"
Repository = "https://github.com/najeebjebreel/lira_analysis"
"Bug Tracker" = "https://github.com/najeebjebreel/lira_analysis/issues"

[tool.setuptools.packages.find]
include = ["utils", "utils.*", "attacks", "attacks.*", "comprehensive_analysis", "comprehensive_analysis.*"]
exclude = ["build*", "data*", "experiments*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
  | experiments
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["utils", "attacks", "comprehensive_analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
